+++ docs/start/changelog.rst
@@ -3,6 +3,11 @@
 ChangeLog
 =========
 
+0.23.1 - TBD
+    * Fixed how tagged and addressable are determined so that they are the
+      correct values based on target when that is set after the packet has been
+      created.
+
 0.23.0 - 12 October 2019
    * Added large_font and speed options to the tile_marquee animation which
      allows a 16x16 font across two tile sets.